cyphron/simulator/requirements.txt
pydantic>=2.7
faker>=24.0
google-cloud-pubsub>=2.21
python-dotenv>=1.0

